backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
cryptography==42.0.8
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
